refactor: move classes to appropriate views
follow django best practices
