+++ app/backend/case/views/case_count_view.py
