+++ app/backend/case/views/patient_case_view.py
